+++ python/pathway/xpacks/llm/_parser_utils.py
@@ -1,8 +1,14 @@
 # Copyright © 2024 Pathway
 
+import asyncio
 import base64
 import io
 import logging
+import os
+import subprocess
+import tempfile
+from collections.abc import Callable
+from typing import Literal
 
 import PIL.Image
 from pydantic import BaseModel
@@ -13,13 +19,6 @@ from pathway.optional_import import optional_imports
 from pathway.xpacks.llm.constants import DEFAULT_VISION_MODEL
 
 logger = logging.getLogger(__name__)
-
-
-def img_to_b64(img: PIL.Image.Image) -> str:
-    buffer = io.BytesIO()
-    img.save(buffer, format="PNG")
-    img_bytes = buffer.getbuffer()
-    return base64.b64encode(img_bytes).decode("utf-8")
 
 
 def maybe_downscale(
@@ -46,7 +45,122 @@ def maybe_downscale(
     return img
 
 
-async def parse(
+def img_to_b64(img: PIL.Image.Image) -> str:
+    buffer = io.BytesIO()
+    img.save(buffer, format="PNG")
+    img_bytes = buffer.getbuffer()
+    return base64.b64encode(img_bytes).decode("utf-8")
+
+
+async def parse_images(
+    images: list[PIL.Image.Image],
+    llm: pw.UDF,
+    parse_prompt: str,
+    *,
+    run_mode: Literal["sequential", "parallel"] = "parallel",
+    parse_details: bool = False,
+    detail_parse_schema: type[BaseModel] | None = None,
+    parse_fn: Callable,
+    parse_image_details_fn: Callable | None,
+) -> tuple[list[str], list[BaseModel]]:
+    """
+    Parse images and optional Pydantic model with a multi-modal LLM.
+    `parse_prompt` will be only used for the regular parsing.
+
+    Args:
+        images: Image list to be parsed. Images are expected to be `PIL.Image.Image`.
+        llm: LLM model to be used for parsing. Needs to support image input.
+        parse_details: Whether to make second LLM call to parse specific Pydantic
+            model from the image.
+        run_mode: Mode of execution,
+            either ``"sequential"`` or ``"parallel"``. Default is ``"parallel"``.
+            ``"parallel"`` mode is suggested for speed, but if timeouts or memory usage in local LLMs are concern,
+            ``"sequential"`` may be better.
+        parse_details: Whether a schema should be parsed.
+        detail_parse_schema: Pydantic model for schema to be parsed.
+        parse_fn: Awaitable image parsing function.
+        parse_image_details_fn: Awaitable image schema parsing function.
+
+    """
+    logger.info("`parse_images` converting images to base64.")
+
+    b64_images = [img_to_b64(image) for image in images]
+
+    return await _parse_b64_images(
+        b64_images,
+        llm,
+        parse_prompt,
+        run_mode=run_mode,
+        parse_details=parse_details,
+        detail_parse_schema=detail_parse_schema,
+        parse_fn=parse_fn,
+        parse_image_details_fn=parse_image_details_fn,
+    )
+
+
+async def _parse_b64_images(
+    b64_images: list[str],
+    llm: pw.UDF,
+    parse_prompt: str,
+    *,
+    run_mode: Literal["sequential", "parallel"],
+    parse_details: bool,
+    detail_parse_schema: type[BaseModel] | None,
+    parse_fn: Callable,
+    parse_image_details_fn: Callable | None,
+) -> tuple[list[str], list[BaseModel]]:
+    total_images = len(b64_images)
+
+    if parse_details:
+        assert detail_parse_schema is not None and issubclass(
+            detail_parse_schema, BaseModel
+        ), "`detail_parse_schema` must be valid Pydantic Model class when `parse_details` is True"
+
+    logger.info(f"`parse_images` parsing descriptions for {total_images} images.")
+
+    parsed_details: list[BaseModel] = []
+
+    if run_mode == "sequential":
+        parsed_content = []
+
+        for img in b64_images:
+            parsed_txt = await parse_fn(img, llm, parse_prompt)
+            parsed_content.append(parsed_txt)
+
+        if parse_details:
+            assert parse_image_details_fn is not None
+            parsed_details = []
+            for img in b64_images:
+                parsed_detail = await parse_image_details_fn(
+                    img,
+                    parse_schema=detail_parse_schema,
+                )
+                parsed_details.append(parsed_detail)
+
+    else:
+        parse_tasks = [parse_fn(img, llm, parse_prompt) for img in b64_images]
+
+        if parse_details:
+            assert parse_image_details_fn is not None
+            detail_tasks = [
+                parse_image_details_fn(
+                    img,
+                    parse_schema=detail_parse_schema,
+                )
+                for img in b64_images
+            ]
+        else:
+            detail_tasks = []
+
+        results = await asyncio.gather(*parse_tasks, *detail_tasks)
+
+        parsed_content = results[: len(b64_images)]
+        parsed_details = results[len(b64_images) :]
+
+    return parsed_content, parsed_details
+
+
+async def parse_image(
     b_64_img,
     llm: pw.UDF,
     prompt: str,
@@ -133,3 +247,36 @@ async def parse_image_details(
     )
 
     return user_info
+
+
+def _convert_pptx_to_pdf(contents: bytes) -> bytes:
+    with tempfile.NamedTemporaryFile(suffix=".pptx", delete=False) as pptx_temp:
+        pptx_temp.write(contents)
+        pptx_temp_path = pptx_temp.name
+
+    pdf_temp_path = pptx_temp_path.replace(".pptx", ".pdf").split(os.path.sep)[-1]
+
+    try:
+        result = subprocess.run(
+            ["soffice", "--headless", "--convert-to", "pdf", pptx_temp_path],
+            check=True,
+            capture_output=True,
+            text=True,
+        )
+
+        logger.info(f"`_convert_pptx_to_pdf` result: {str(result)}")
+
+        with open(pdf_temp_path, "rb") as pdf_temp:
+            pdf_contents = pdf_temp.read()
+
+    except FileNotFoundError:
+        raise Exception(
+            "`LibreOffice` is not installed or `soffice` command is not found. Please install LibreOffice."
+        )
+
+    finally:
+        os.remove(pptx_temp_path)
+        if os.path.exists(pdf_temp_path):
+            os.remove(pdf_temp_path)
+
+    return pdf_contents